Introduce 'blank state' play button asset for gif.
Some tweaks to CSS also. The UX is still not perfect here. There
is a noticeable delay after clicking the gif till user feedback.
I blieve this is caused by the main thread blocking as the frames
for the gif are generated, so I think the solution would be to
send this work into a web worker thread.

I might come back to this or I might just decide the status quo
will have to do. One unfortunate aspect is that web workers must
serialise and copy all data so we'd take a hit transferring all
the frames etc.